Solved View - View model Sync issue
Found a solution for maintaining the nested view models in sync with
each other and the view itself.  Solution uses KO pub-sub to pass
messages and handle data and events accordingly.  Very useful.  Now that
I understand the good and the bad, how it works and what to expect,  I
actually like KO!